[tf.data] Fixing type mismatch, which could lead to loss of precision sanitizer errors.
PiperOrigin-RevId: 322434758

+++ tensorflow/core/framework/model.cc
@@ -304,7 +304,7 @@ class AsyncInterleaveMany : public Node {
 
 class KnownRatio : public Node {
  public:
-  KnownRatio(Node::Args args, int64 ratio) : Node(args), ratio_(ratio) {}
+  KnownRatio(Node::Args args, double ratio) : Node(args), ratio_(ratio) {}
 
   virtual ~KnownRatio() {}
